+++ kanban/frontend/src/app/features/boards/components/boards-list/boards-list.html
