+++ kanban/frontend/src/app/features/boards/models/board.model.ts
