+++ kanban/frontend/src/app/features/boards/services/boards.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Board } from './board.model';
+import { Board } from '../models/board.model';
 import { Observable } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
